feat: add InternalMoveGeneralBody
