+++ packages/apps/stock/src/components/organisms/index.js
@@ -44,3 +44,4 @@ export {default as StockCorrectionHeader} from './StockCorrectionHeader/StockCor
 export {default as StockCorrectionBody} from './StockCorrectionBody/StockCorrectionBody';
 export {default as StockCorrectionFixedItems} from './StockCorrectionFixedItems/StockCorrectionFixedItems';
 export {default as InternalMoveGeneralFixedItems} from './InternalMoveGeneralFixedItems/InternalMoveGeneralFixedItems';
+export {default as InternalMoveGeneralBody} from './InternalMoveGeneralBody/InternalMoveGeneralBody';
