+++ packages/apps/stock/src/screens/internalMoves/InternalMoveDetailsGeneralScreen.js
@@ -17,17 +17,7 @@
  */
 
 import React, {useEffect} from 'react';
-import {StyleSheet} from 'react-native';
-import {
-  HeaderContainer,
-  Icon,
-  MovementIndicationCard,
-  Screen,
-  ScrollView,
-  ViewAllContainer,
-  useThemeColor,
-  NotesCard,
-} from '@axelor/aos-mobile-ui';
+import {HeaderContainer, Screen, ScrollView} from '@axelor/aos-mobile-ui';
 import {
   useDispatch,
   useSelector,
@@ -35,24 +25,21 @@ import {
   HeaderOptionsMenu,
 } from '@axelor/aos-mobile-core';
 import {
+  InternalMoveGeneralBody,
   InternalMoveGeneralFixedItems,
-  InternalMoveLineCard,
   StockMoveHeader,
 } from '../../components';
 import {fetchInternalMoveLines} from '../../features/internalMoveLineSlice';
 import {getRacks} from '../../features/racksListSlice';
 import StockMove from '../../types/stock-move';
-import {showLine} from '../../utils/line-navigation';
 
 const InternalMoveDetailsGeneralScreen = ({navigation, route}) => {
   const internalMove = route.params.internalMove;
   const {loadingIMLines, internalMoveLineList} = useSelector(
     state => state.internalMoveLine,
   );
-  const {loadingRacks, racksList} = useSelector(state => state.rack);
   const {mobileSettings} = useSelector(state => state.config);
   const I18n = useTranslator();
-  const Colors = useThemeColor();
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -71,24 +58,6 @@ const InternalMoveDetailsGeneralScreen = ({navigation, route}) => {
       }),
     );
   }, [dispatch, internalMove, internalMoveLineList]);
-
-  const handleViewAll = () => {
-    navigation.navigate('InternalMoveLineListScreen', {
-      internalMove: internalMove,
-    });
-  };
-
-  const handleShowLine = item => {
-    showLine({
-      item: {name: 'internalMove', data: internalMove},
-      itemLine: {name: 'internalMoveLine', data: item},
-      lineDetailsScreen: 'InternalMoveLineDetailsScreen',
-      selectTrackingScreen: 'InternalMoveSelectTrackingScreen',
-      selectProductScreen: 'InternalMoveSelectProductScreen',
-      productKey: 'stockProduct',
-      navigation,
-    });
-  };
 
   React.useLayoutEffect(() => {
     navigation.setOptions({
@@ -128,59 +97,13 @@ const InternalMoveDetailsGeneralScreen = ({navigation, route}) => {
         }
       />
       <ScrollView>
-        <MovementIndicationCard
-          titleTop={internalMove.fromStockLocation.name}
-          iconTop={
-            <Icon name="warehouse" color={Colors.primaryColor.background} />
-          }
-          titleDown={internalMove.toStockLocation.name}
-          iconDown={
-            <Icon name="warehouse" color={Colors.primaryColor.background} />
-          }
-        />
-        <ViewAllContainer
-          data={internalMoveLineList}
-          renderFirstTwoItems={(item, index) => (
-            <InternalMoveLineCard
-              style={styles.item}
-              productName={item.product?.fullName}
-              internalMoveStatus={internalMove.statusSelect}
-              availability={
-                item.availableStatusSelect != null
-                  ? item.availableStatusSelect
-                  : null
-              }
-              locker={
-                !loadingRacks && racksList != null && racksList[index] != null
-                  ? racksList[index][0]?.rack
-                  : ''
-              }
-              trackingNumber={item.trackingNumber?.trackingNumberSeq}
-              expectedQty={item.qty}
-              movedQty={item.realQty}
-              onPress={() => handleShowLine(item)}
-            />
-          )}
-          onViewPress={handleViewAll}
-        />
-        <NotesCard
-          title={I18n.t('Stock_NotesOnPreparation')}
-          data={internalMove.pickingOrderComments}
-        />
-        <NotesCard
-          title={I18n.t('Stock_NotesOnStockMove')}
-          data={internalMove.note}
+        <InternalMoveGeneralBody
+          internalMove={internalMove}
+          navigation={navigation}
         />
       </ScrollView>
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  item: {
-    marginHorizontal: 1,
-    marginVertical: 4,
-  },
-});
 
 export default InternalMoveDetailsGeneralScreen;
